+++ src/command_list/cmd_list_free.c
@@ -1,0 +1,34 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   cmd_list_free.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/09 10:21:52 by Philip            #+#    #+#             */
+/*   Updated: 2024/04/09 10:29:37 by Philip           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cmd_list.h"
+#include "../string_array.h"
+#include <stdlib.h>
+
+void	cmd_list_free(t_cmd_list **list)
+{
+	t_cmd_list	*node;
+	t_cmd_list	*next_node;
+
+	if (!list || !(*list))
+		return ;
+	node = *list;
+	while (node)
+	{
+		next_node = node->next;
+		free_string_array_and_null(&node->cmd_argv);
+		free_string_array_and_null(&node->redirects);
+		free(node);
+		node = next_node;
+	}
+	*list = NULL;
+}
