+++ src/expand_arguments.c
@@ -6,11 +6,11 @@
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 17:07:24 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/09 10:17:45 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/09 10:27:26 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cmd_list.h"
+#include "command_list/cmd_list.h"
 #include "environment_variables/env.h"
 #include "character_list/char_list.h"
 #include "free_and_null.h"
